Update Integration.ipynb
Removing this link since it is not found: http://cbdm-01.zdv.uni-mainz.de/~jibnsale/teaching/motif_analysis_exercises.html\n. It appears to be a link to a professor's personal teaching site but nothing is found at the head domain either. Removing it does not impact the notebook functionality at all since it is in the "Further Reading" section.